refactor(lowy): cross-validate — rename RowIcons → RowPips
`RowIcons.tsx` named the file by location ("the row icons"). Lowy
cross-validation flagged this as functional decomposition: the
three exports (`AgentSlot`, `PrPip`, `SubCountCell`) each
encapsulate an independent volatility axis, but the filename told
you nothing about which axes lived inside.

Renames to `RowPips.tsx` — "pip" is a noun for the thing (a
small presence indicator), which signals what the file contains.
Adds a header comment naming the three volatility axes so a
maintainer searching for "how is the PR pip rendered" finds the
right module from `git grep` alone.

Splitting into three files was the stricter Lowy move, but the
file is ~120 lines and the three pieces are consumed identically
by both dock surfaces — the location grouping is honest until one
of the axes diverges further. The comment names that split-out
condition explicitly.
