+++ packages/client/src/canvas/dock/SubCountChip.tsx
@@ -10,7 +10,7 @@
  *
  *  `testId` is required (not optional with a default) so each call
  *  site is testable by a stable id. Both desktop and mobile dock rows
- *  render this via the shared `RowIcons` component, which passes
+ *  render this via the shared `RowPips` `SubCountCell`, which passes
  *  `testId="dock-sub-count"`. */
 
 import type { Component } from "solid-js";
